requirements.txt
allure-pytest==2.8.18
allure-python-commons==2.8.18
atomicwrites==1.4.0
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
colorlog==4.2.1
har2case==0.3.1
HttpRunner==1.5.8
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.5.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
PyMySQL==0.10.1
pyparsing==2.4.7
pytest==6.0.2
pytest-html==2.1.1
pytest-metadata==1.10.0
PyYAML==5.3.1
requests==2.24.0
requests-toolbelt==0.9.1
six==1.15.0
toml==0.10.1
urllib3==1.25.10
zipp==3.1.0
